requirements.txt
amqp==1.4.9
anyjson==0.3.3
awscli==1.11.13
awsebcli==3.8.9
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.23
blessed==1.9.5
boto==2.43.0
botocore==1.4.87
celery==3.1.25
cement==2.8.2
cffi==1.9.1
colorama==0.3.7
configparser==3.5.0
contextlib2==0.5.4
cryptography==1.7.1
decorator==4.0.10
Django==1.9
django-activity-stream==0.6.3
django-braces==1.10.0
django-celery==3.1.17
django-constance==1.3.4
django-contrib-comments==1.7.3
django-cors-headers==1.3.1
django-extensions==1.7.4
django-filter==0.15.3
django-filters==0.2.1
django-fsm==2.4.0
django-fsm-admin==1.2.4
django-jsonfield==1.0.1
django-oauth-toolkit==0.11.0
django-picklefield==0.3.2
django-rest-framework-social-oauth2==1.0.4
django-storages==1.5.1
django-versatileimagefield==1.6
djangorestframework==3.4.0
djangorestframework-gis==0.10.1
djangorestframework-jwt==1.9.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
enum34==1.1.6
Fabric==1.13.1
flake8==3.0.4
futures==3.0.5
http==0.2
idna==2.2
ipaddress==1.0.17
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
jmespath==0.9.0
kombu==3.0.37
mccabe==0.5.2
newrelic==2.78.0.57
oauthlib==1.1.2
paramiko==2.1.1
pathlib2==2.1.0
pathspec==0.5.0
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.8
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.9
pycodestyle==2.0.0
pycparser==2.17
pyflakes==1.2.3
Pygments==2.1.3
PyJWT==1.4.2
python-dateutil==2.6.0
python-openid==2.2.5
python-social-auth==0.2.21
pytz==2016.10
PyYAML==3.12
raven==5.32.0
requests==2.9.1
requests-oauthlib==0.7.0
rsa==3.4.2
s3transfer==0.1.9
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.10.0
stevedore==1.18.0
tabulate==0.7.5
texttable==0.8.7
traitlets==4.3.1
vine==1.1.3
virtualenv==15.0.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wcwidth==0.1.7
websocket-client==0.40.0
